requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
discord.py==1.7.3
Flask==2.1.2
future==0.18.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
multidict==6.0.2
python-dotenv==0.20.0
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
